// lib/typings/types.ts
export enum ProviderType {
  bitbucket_cloud = "bitbucket_cloud",
  github_com = "github_com",
  ghe = "ghe",
  bitbucket = "bitbucket",
  gitlab = "gitlab"
}

/* Enum for OwnerType */
export enum OwnerType {
  user = "user",
  organization = "organization"
}
// ...
  export type Repo = {
    __typename?: "Repo";
    branches?: Branches[] | null;
  };

  export type Branches = {
    __typename?: "Branch";
    commit?: Commit | null;
  };
// ...
  export type ChatId = {
    __typename?: "ChatId";
    userId?: string | null;
    screenName?: string | null;
    person?: Person | null;
  };

  export type Person = PersonFields.Fragment;
// ...
  export type Org = {
    __typename?: "Org";
    provider?: Provider | null;
  };

  export type Provider = {
    __typename?: "GitHubProvider";
    providerId?: string | null;
  };
// ...
  export type Team = {
    __typename?: "ChatTeam";
    id?: string | null;
  };
// ...
  export type Emails = {
    __typename?: "Email";
    address?: string | null;
  };
// ...
export namespace CoreRepoFieldsAndChannels {
  export type Fragment = {
    __typename?: "Repo";
    owner?: string | null;
    name?: string | null;
    org?: Org | null;
    channels?: Channels[] | null;
    defaultBranch?: string | null;
  };

  export type Org = {
    __typename?: "Org";
    owner?: string | null;
    ownerType?: OwnerType | null;
    provider?: Provider | null;
  };

  export type Provider = {
    __typename?: "GitHubProvider";
    providerId?: string | null;
    providerType?: ProviderType | null;
    apiUrl?: string | null;
    url?: string | null;
  };

  export type Channels = {
    __typename?: "ChatChannel";
    team?: Team | null;
    name?: string | null;
    id?: string | null;
  };

  export type Team = {
    __typename?: "ChatTeam";
    id?: string | null;
  };
}

export namespace PersonFields {
  export type Fragment = {
    __typename?: "Person";
    forename?: string | null;
    surname?: string | null;
    name?: string | null;
    emails?: Emails[] | null;
    gitHubId?: GitHubId | null;
    chatId?: ChatId | null;
  };

  export type Emails = {
    __typename?: "Email";
    address?: string | null;
  };

  export type GitHubId = {
    __typename?: "GitHubId";
    login?: string | null;
  };

  export type ChatId = {
    __typename?: "ChatId";
    screenName?: string | null;
  };
}